changed a feww more variables to double, tweaked movement speed and mouse sensitivity

+++ srcs/render/raycast-3d.c
@@ -1,15 +1,15 @@
 #include "cub3d.h"
 
-static float	fix_fisheye(t_game *g, float angle, float dist)
+static double	fix_fisheye(t_game *g, double angle, double dist)
 {
-	float	fisheye;
-	float	new_dist;
+	double	fisheye;
+	double	new_dist;
 
 	fisheye = g->p.pa - angle;
 	if (fisheye < 0)
-		fisheye += 2 * M_PI;
-	if (fisheye > 2 * M_PI)
-		fisheye -= 2 * M_PI;
+		fisheye += 2.0 * M_PI;
+	if (fisheye > 2.0 * M_PI)
+		fisheye -= 2.0 * M_PI;
 	new_dist = dist * cos(fisheye);
 	return (new_dist);
 }
@@ -22,16 +22,16 @@ static void	draw_3d(t_game *g, t_vecd start, t_vecd end)
 	double	ray_w_end;
 
 	ray_w = start.x;
-	ray_w_end = start.x + 1.0f;
+	ray_w_end = start.x + 1.0;
 	while (ray_w < ray_w_end)
 	{
 		ray_h = start.y;
 		while(ray_h <= end.y)
 		{
 			my_pp(g->bg, round(ray_w), round(ray_h), PURPLE);
-			ray_h += 1.0f;
+			ray_h += 1.0;
 		}
-		ray_w += 1.0f;
+		ray_w += 1.0;
 	}
 }
 
@@ -50,7 +50,7 @@ void	cast_3d(t_game *g, double dist, int r, double angle)
 	lineH = ((double)g->winsize.y * (double)CELL_SIZE) / (double)dist;
 	if (lineH > (double)g->winsize.y)
 		lineH = (double)g->winsize.y;
-	lineOff = ((double)g->winsize.y / 2.0f) - lineH / 2.0f;
+	lineOff = ((double)g->winsize.y / 2.0) - lineH / 2.0;
 	start = (t_vecd){.x = (double)r, .y = lineOff};
 	end = (t_vecd){.x = (double)r, .y = lineOff + lineH};
 	draw_3d(g, start, end);
